+++ src/main/java/com/perifacode/gerenciador/adapter/controller/ExceptionHandlerController.java
@@ -1,0 +1,24 @@
+package com.perifacode.gerenciador.adapter.controller;
+
+import javax.validation.ConstraintViolationException;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.ControllerAdvice;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.context.request.WebRequest;
+import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler;
+
+@ControllerAdvice
+public class ExceptionHandlerController extends ResponseEntityExceptionHandler {
+
+  @ExceptionHandler(
+      value = {DataIntegrityViolationException.class})
+  protected ResponseEntity<Object> handleConflict(
+      DataIntegrityViolationException ex, WebRequest request) {
+    String bodyOfResponse = "Falha ao efetuar a requisição, conflito";
+    return handleExceptionInternal(ex, bodyOfResponse,
+        new HttpHeaders(), HttpStatus.CONFLICT, request);
+  }
+}
