+++ src/main/java/com/perifacode/gerenciador/entity/TipoIniciativa.java
@@ -1,6 +1,6 @@
 package com.perifacode.gerenciador.entity;
 
-import java.time.LocalDateTime;
+import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
@@ -8,21 +8,17 @@ import javax.persistence.Id;
 import javax.validation.constraints.NotBlank;
 
 @Entity
-public class TipoIniciativa {
+public class TipoIniciativa extends BaseEntity {
 
   public TipoIniciativa(){}
 
   public TipoIniciativa(long id, @NotBlank String codigo,
                         @NotBlank String titulo,
-                        @NotBlank String descricao,
-                        @NotBlank LocalDateTime dataCadastro,
-                        @NotBlank LocalDateTime dataAtualizacao, boolean ativo) {
+                        @NotBlank String descricao, boolean ativo) {
     this.id = id;
     this.codigo = codigo;
     this.titulo = titulo;
     this.descricao = descricao;
-    this.dataCadastro = dataCadastro;
-    this.dataAtualizacao = dataAtualizacao;
     this.ativo = ativo;
   }
 
@@ -31,19 +27,15 @@ public class TipoIniciativa {
   private long id;
 
   @NotBlank
+  @Column(unique = true)
   private String codigo;
 
   @NotBlank
+  @Column(unique = true)
   private String titulo;
 
   @NotBlank
   private String descricao;
-
-  @NotBlank
-  private LocalDateTime dataCadastro;
-
-  @NotBlank
-  private LocalDateTime dataAtualizacao;
 
   private boolean ativo;
 
@@ -77,22 +69,6 @@ public class TipoIniciativa {
 
   public void setDescricao(String descricao) {
     this.descricao = descricao;
-  }
-
-  public LocalDateTime getDataCadastro() {
-    return dataCadastro;
-  }
-
-  public void setDataCadastro(LocalDateTime dataCadastro) {
-    this.dataCadastro = dataCadastro;
-  }
-
-  public LocalDateTime getDataAtualizacao() {
-    return dataAtualizacao;
-  }
-
-  public void setDataAtualizacao(LocalDateTime dataAtualizacao) {
-    this.dataAtualizacao = dataAtualizacao;
   }
 
   public boolean isAtivo() {
