+++ src/main/resources/application.yml
@@ -1,10 +1,12 @@
 spring:
+  jackson:
+    property-naming-strategy: SNAKE_CASE
   h2:
     console:
       enabled: true
       path: /h2
   datasource:
-    url: jdbc:h2:file:~/managerDb
+    url: jdbc:h2:mem:~/managerDb
     username: admin
     password: admin
     driver-class-name: org.h2.Driver
